refactor: standalone string format and game related utils
